moved UpdateWithdrawalBalance
